// app-cars/src/app/source/car-interface/car-interface.ts
export interface CarI {
    name: string;
    year: string;
    model: string;
    description: string;
    fuel: string;
    doors: number;
    type: string;
    color: string;
    kms: number;
    price: number;
    reserved: boolean;
    owner: any[];

    dgt_reg: string;
    accidentFree: boolean;
    shop_name: string;
    shop_location: string;
    image: string[];
    gallery: string[];
    id: string;
    shop_country: any;
    images: string[];
}